requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
exceptiongroup==1.2.2
execnet==2.1.1
greenlet==3.2.1
idna==3.10
iniconfig==2.1.0
packaging==25.0
pillow==11.2.1
playwright==1.52.0
pluggy==1.5.0
pyee==13.0.0
pytesseract==0.3.13
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-base-url==2.1.0
pytest-playwright==0.7.0
pytest-xdist==3.6.1
python-dotenv==1.1.0
python-slugify==8.0.4
requests==2.32.3
text-unidecode==1.3
tomli==2.2.1
typing_extensions==4.13.2
urllib3==2.4.0
